Refactored the code and implemented the dropdown
