+++ console-extensions.json
@@ -20,12 +20,30 @@
       "href": "/catalog?catalogType=dbaas-new",
       "label": "Connected Database",
       "description": "Select a cloud database provider to connect",
-      "icon": { "$codeRef": "constants.dbaasServicesIcon" },
+      "icon": {
+        "$codeRef": "constants.dbaasServicesIcon"
+      },
       "accessReview": [
-        { "group": "apps.openshift.io", "resource": "deploymentconfigs", "verb": "create" },
-        { "group": "route.openshift.io", "resource": "routes", "verb": "create" },
-        { "group": "", "resource": "secrets", "verb": "create" },
-        { "group": "", "resource": "services", "verb": "create" }
+        {
+          "group": "apps.openshift.io",
+          "resource": "deploymentconfigs",
+          "verb": "create"
+        },
+        {
+          "group": "route.openshift.io",
+          "resource": "routes",
+          "verb": "create"
+        },
+        {
+          "group": "",
+          "resource": "secrets",
+          "verb": "create"
+        },
+        {
+          "group": "",
+          "resource": "services",
+          "verb": "create"
+        }
       ]
     }
   },
@@ -46,7 +64,9 @@
     "properties": {
       "catalogId": "dev-catalog",
       "type": "dbaas-new",
-      "provider": { "$codeRef": "catalog.dbaasProvider" }
+      "provider": {
+        "$codeRef": "catalog.dbaasProvider"
+      }
     },
     "flags": {
       "required": []
@@ -60,7 +80,9 @@
         "version": "v1alpha1",
         "kind": "DBaaSInventory"
       },
-      "component": { "$codeRef": "dbaasImportPage" }
+      "component": {
+        "$codeRef": "dbaasImportPage"
+      }
     }
   },
   {
@@ -89,7 +111,7 @@
     "properties": {
       "exact": true,
       "path": [
-        "/rhoda-admin-dashboard"
+        "/k8s/ns/:ns/rhoda-admin-dashboard"
       ],
       "component": {
         "$codeRef": "adminDashboard"
@@ -100,35 +122,11 @@
     "type": "console.navigation/href",
     "properties": {
       "id": "rhoda-admin-dashboard",
+      "prefixNamespaced": true,
       "perspective": "admin",
       "section": "data-services",
       "name": "Database Access",
       "href": "/rhoda-admin-dashboard"
     }
   }
-  // {
-  //   "type": "console.page/route",
-  //   "properties": {
-  //     "exact": true,
-  //     "path": [
-  //       "/k8s/ns/:ns/rhoda-admin-dashboard"
-  //     ],
-  //     "component": {
-  //       "$codeRef": "adminDashboard"
-  //     }
-  //   }
-  // },
-  // {
-  //   "type": "console.navigation/href",
-  //   "properties": {
-  //     "id": "rhoda-admin-dashboard",
-  //     "perspective": "admin",
-  //     "name": "Data Services",
-  //     "prefixNamespaced": true,
-  //     "href": "/rhoda-admin-dashboard"
-  //   },
-  //   "flags": {
-  //     "required": ["OPENSHIFT"]
-  //   }
-  // }
 ]
