+++ src/const.ts
@@ -8,12 +8,8 @@ export const DBaaSOperatorName = 'dbaas-operator'
 
 // Currently placeholders
 export const dbaasServicesIcon = `data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' enable-background='new 0 0 66 66' viewBox='0 0 66 66'%3E%3Cpath d='M28.2888 34.4121h-4.4478c-.5522 0-1 .4473-1 1s.4478 1 1 1h4.4478c.5522 0 1-.4473 1-1S28.8411 34.4121 28.2888 34.4121zM41.9519 34.4121h-4.4478c-.5522 0-1 .4473-1 1s.4478 1 1 1h4.4478c.5522 0 1-.4473 1-1S42.5042 34.4121 41.9519 34.4121z'/%3E%3Ccircle cx='33' cy='35.412' r='1.033'/%3E%3Cpath d='M49.3796,52.294c-0.5522,0-1,0.4473-1,1V64c0,0.5527,0.4478,1,1,1h10.7056c0.5522,0,1-0.4473,1-1V53.294 c0-0.5527-0.4478-1-1-1h-4.4658v-5.9932c0-0.5527-0.4478-1-1-1H37.2229c-0.3673-1.3762-1.459-2.452-2.8481-2.7869v-2.1438h9.9614 c1.3442,0,2.4375-1.0938,2.4375-2.4375v-1.5244h9.2856c0.5522,0,1-0.4473,1-1v-5.7598h2.9429c0.5522,0,1-0.4473,1-1v-7.8799 c0-0.5527-0.4478-1-1-1h-7.8789c-0.5522,0-1,0.4473-1,1v7.8799c0,0.5527,0.4478,1,1,1h2.936v4.7598h-8.2856v-1.5156 c0-1.3438-1.0933-2.4375-2.4375-2.4375h-9.166l4.3896-4.3896c0.3906-0.3906,0.3906-1.0234,0-1.4141s-1.0234-0.3906-1.4141,0 l-4.1533,4.1533v-6.3203h9.3018c3.3569,0,6.0879-2.7314,6.0879-6.0879c0-3.3154-2.6572-6.0215-5.9502-6.0918 C44.5022,5.7263,41.0228,1,36.01,1c-3.4453,0-6.3955,2.293-7.3057,5.5039c-2.7204-0.6306-5.5206,1.0497-6.2461,3.7793 c-2.6831,0.1553-5.5591,2.6221-5.5591,6.1133c0,3.3564,2.731,6.0879,6.0879,6.0879h9.0054v6.1704l-4.0034-4.0034 c-0.3906-0.3906-1.0234-0.3906-1.4141,0s-0.3906,1.0234,0,1.4141l4.3896,4.3896h-9.3008c-1.3438,0-2.437,1.0938-2.437,2.4375 v1.5156h-8.2729v-4.7598h2.9365c0.5522,0,1-0.4473,1-1v-7.8799c0-0.5527-0.4478-1-1-1H6.011c-0.5522,0-1,0.4473-1,1v7.8799 c0,0.5527,0.4478,1,1,1h2.9429v5.7598c0,0.5527,0.4478,1,1,1h9.2729v1.5244c0,1.3438,1.0933,2.4375,2.437,2.4375h10.7109v2.1968 c-1.3128,0.3745-2.3187,1.4094-2.6644,2.7339H10.9382c-0.5522,0-1,0.4473-1,1v5.9932H5.9148c-0.5522,0-1,0.4473-1,1V64 c0,0.5527,0.4478,1,1,1h10.7061c0.5522,0,1-0.4473,1-1V53.294c0-0.5527-0.4478-1-1-1h-4.6826v-4.9932h17.7789 c0.3541,1.301,1.3656,2.3386,2.6577,2.7195v2.2737h-4.8125c-0.5522,0-1,0.4473-1,1V64c0,0.5527,0.4478,1,1,1h10.7056 c0.5522,0,1-0.4473,1-1V53.294c0-0.5527-0.4478-1-1-1h-3.8931v-2.2148c1.3884-0.3339,2.4794-1.4063,2.8474-2.7783h16.3972v4.9932 H49.3796z M53.1233,27.6485v-5.8799h5.8789v5.8799H53.1233z M7.011,27.6485v-5.8799h5.8794v5.8799H7.011z M15.6208,54.294V63 H6.9148V54.294H15.6208z M22.9871,20.4844c-2.2539,0-4.0879-1.834-4.0879-4.0879c0-2.5156,2.2651-4.1201,3.8247-4.1201 c0.4297,0,0.7464,0.1782,1.1836-0.1602c0.2246-0.1748,0.3643-0.4365,0.3843-0.7207c0.1619-2.3235,2.6884-3.7122,4.7075-2.6689 c0.6275,0.3242,1.3739-0.0864,1.4521-0.7764C30.7703,5.128,33.1599,3,36.01,3c4.266,0,6.8109,4.5071,5.043,8.041 c-0.1768,0.3535-0.1289,0.7773,0.1221,1.083c0.5545,0.6731,1.0922,0.1797,2.125,0.1797c2.251,0,4.082,1.8359,4.082,4.0928 c0,2.2539-1.834,4.0879-4.0879,4.0879H22.9871z M21.6638,38.3701c-0.2412,0-0.437-0.1963-0.437-0.4375v-5.04 c0-0.2412,0.1958-0.4375,0.437-0.4375h22.6724c0.2412,0,0.4375,0.1963,0.4375,0.4375v5.04c0,0.2412-0.1963,0.4375-0.4375,0.4375 H21.6638z M37.2678,54.294V63h-8.7056V54.294H37.2678z M33.51,48.1963c-0.0361-0.0049-0.0728-0.0078-0.1094-0.0088 c-1.0107-0.0254-1.833-0.8721-1.833-1.8867c0-1.0195,0.8223-1.8701,1.833-1.8955c0.0366-0.001,0.0732-0.0039,0.1094-0.0088 c1.0293,0.0244,1.8589,0.8691,1.8589,1.9043C35.3689,47.3311,34.5393,48.1719,33.51,48.1963z M59.0852,54.294V63h-8.7056V54.294 H59.0852z'/%3E%3C/svg%3E`
-export const mongoDBIcon = `data:image/svg+xml,%3Csvg height='64' viewBox='0 0 32 32' width='64' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='m15.9.087.854 1.604c.192.296.4.558.645.802.715.715 1.394 1.464 2.004 2.266 1.447 1.9 2.423 4.01 3.12 6.292.418 1.394.645 2.824.662 4.27.07 4.323-1.412 8.035-4.4 11.12-.488.488-1.01.94-1.57 1.342-.296 0-.436-.227-.558-.436-.227-.383-.366-.82-.436-1.255-.105-.523-.174-1.046-.14-1.586v-.244c-.024-.052-.285-24.052-.181-24.175z' fill='%23599636'/%3E%3Cpath d='m15.9.034c-.035-.07-.07-.017-.105.017.017.35-.105.662-.296.96-.21.296-.488.523-.767.767-1.55 1.342-2.77 2.963-3.747 4.776-1.3 2.44-1.97 5.055-2.16 7.808-.087.993.314 4.497.627 5.508.854 2.684 2.388 4.933 4.375 6.885.488.47 1.01.906 1.55 1.325.157 0 .174-.14.21-.244a4.78 4.78 0 0 0 .157-.68l.35-2.614z' fill='%236cac48'/%3E%3Cpath d='m16.754 28.845c.035-.4.227-.732.436-1.063-.21-.087-.366-.26-.488-.453-.105-.174-.192-.383-.26-.575-.244-.732-.296-1.5-.366-2.248v-.453c-.087.07-.105.662-.105.75a17.37 17.37 0 0 1 -.314 2.353c-.052.314-.087.627-.28.906 0 .035 0 .07.017.122.314.924.4 1.865.453 2.824v.35c0 .418-.017.33.33.47.14.052.296.07.436.174.105 0 .122-.087.122-.157l-.052-.575v-1.604c-.017-.28.035-.558.07-.82z' fill='%23c2bfbf'/%3E%3C/svg%3E`
-export const APITokenLengthMinimum = 500
 export const crunchyProviderType = 'crunchy-bridge-registration'
 export const mongoProviderType = 'mongodb-atlas-registration'
-export const adminProviderType = 'rhoda-admin-dashboard'
-export const adminProviderName = 'Admin'
 export const crunchyProviderName = 'Crunchy Bridge Managed PostgreSQL'
 export const mongoProviderName = 'MongoDB Atlas Cloud Database Service'
